PSENet/requirement.txt
mmcv
paddlepaddle-gpu==2.0.1
Polygon3
pyclipper

